Fixing #1206, bot now fails to start on missing git
